A social media program.
An example in which you are completely represented by an abstraction. You are connected to a platform that's available to the world but you can access only a few, others are behind the field. You are represented by your profile, have your friends, your privacy. It all varies with the degrees of abstraction.